api/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt[crypto]
black
channels
channels_redis
django-filter